feat(mining): validate remaining Pearl model artifacts
