+++ docs/development/pearl-model-enablement.md
@@ -13,8 +13,8 @@ for the vanilla Pearl GEMM path.
 | Raw model | Planned Pearl model | Status | Tracking |
 |---|---|---|---|
 | `Qwen/Qwen3.5-9B` | `ScalingIntelligence/Qwen3.5-9B-pearl` | Validated staging | [#316](https://github.com/open-jarvis/OpenJarvis/issues/316) |
-| `Qwen/Qwen3.6-27B` | `pearl-ai/Qwen3.6-27B-pearl` | Planned | [#317](https://github.com/open-jarvis/OpenJarvis/issues/317) |
-| `google/gemma-4-E4B-it` | `pearl-ai/Gemma-4-E4B-it-pearl` | Planned | [#318](https://github.com/open-jarvis/OpenJarvis/issues/318) |
+| `Qwen/Qwen3.6-27B` | `ScalingIntelligence/Qwen3.6-27B-pearl` | Validated staging | [#317](https://github.com/open-jarvis/OpenJarvis/issues/317) |
+| `google/gemma-4-E4B-it` | `ScalingIntelligence/Gemma-4-E4B-it-pearl` | Validated staging | [#318](https://github.com/open-jarvis/OpenJarvis/issues/318) |
 | `google/gemma-4-31B-it` | `ScalingIntelligence/Gemma-4-31B-it-pearl` | Validated staging | [#319](https://github.com/open-jarvis/OpenJarvis/issues/319) |
 
 The current validated models are:
@@ -22,7 +22,9 @@ The current validated models are:
 ```text
 pearl-ai/Llama-3.3-70B-Instruct-pearl
 ScalingIntelligence/Gemma-4-31B-it-pearl
+ScalingIntelligence/Gemma-4-E4B-it-pearl
 ScalingIntelligence/Qwen3.5-9B-pearl
+ScalingIntelligence/Qwen3.6-27B-pearl
 ```
 
 ## Current Validation Findings
@@ -31,16 +33,18 @@ The H100 smoke run validated the default Llama Pearl model end to end through
 `jarvis mine start`, vLLM `/v1/models`, OpenJarvis inference routing, Pearl
 gateway template refresh, and `jarvis mine validate-model`.
 
-The remaining Qwen 3.6 and smaller Gemma target remain planned:
+The upstream Pearl model ids for Qwen 3.6, Gemma E4B, and Gemma 31B remain
+planned because the `pearl-ai/*` artifacts are not available as complete
+validated user-facing repos yet:
 
-- `pearl-ai/Qwen3.6-27B-pearl` and `pearl-ai/Gemma-4-E4B-it-pearl` are not
-  publicly available artifacts yet.
 - `pearl-ai/Gemma-4-31B-it-pearl` exists and selects Pearl mining kernels on
   H100, but vLLM fails during Gemma4 multimodal profiling because the published
   artifact is missing processor/preprocessor metadata required by Transformers.
   A local cache experiment proved the processor can be loaded only after
   injecting metadata from `google/gemma-4-31B-it`; that is not sufficient for
   OpenJarvis promotion because a clean user install would still fail.
+- `pearl-ai/Qwen3.6-27B-pearl` and `pearl-ai/Gemma-4-E4B-it-pearl` remain
+  planned upstream model ids until Pearl publishes/validates them.
 
 PR #323 added a local staging path for original checkpoint conversion and H100
 runtime validation. The validated staging artifacts are private repositories in
@@ -50,7 +54,9 @@ the org/repositories before `jarvis mine inspect-model`, `jarvis mine init`, or
 `jarvis mine start` can fetch them:
 
 - `ScalingIntelligence/Gemma-4-31B-it-pearl`
+- `ScalingIntelligence/Gemma-4-E4B-it-pearl`
 - `ScalingIntelligence/Qwen3.5-9B-pearl`
+- `ScalingIntelligence/Qwen3.6-27B-pearl`
 
 Current validation evidence:
 
@@ -74,6 +80,30 @@ Current validation evidence:
   `/tmp/openjarvis-h100/converted/Qwen3.5-9B-pearl-experimental-validate.json`.
   The published `ScalingIntelligence/Qwen3.5-9B-pearl` repo includes the same
   validation artifact as `openjarvis_validation.json`.
+- `google/gemma-4-E4B-it` converted to
+  `/tmp/openjarvis-h100/converted/Gemma-4-E4B-it-pearl-experimental`.
+  The local artifact passes `jarvis mine inspect-model`, starts Pearl gateway,
+  loads with Gemma4 per-layer embeddings excluded from Pearl quantization,
+  exposes `pearl-ai/Gemma-4-E4B-it-pearl` at `/v1/models`, completes a chat
+  prompt, and passes `jarvis mine validate-model --allow-planned`. Required
+  validation flags: `--language-model-only` and `--skip-mm-profiling`.
+  Validation artifact:
+  `/tmp/openjarvis-h100/converted/Gemma-4-E4B-it-pearl-experimental-validate.json`.
+  The published `ScalingIntelligence/Gemma-4-E4B-it-pearl` repo includes the
+  same validation artifact as `openjarvis_validation.json`.
+- `Qwen/Qwen3.6-27B` converted to
+  `/tmp/openjarvis-h100/converted/Qwen3.6-27B-pearl-experimental`.
+  The local artifact passes `jarvis mine inspect-model`, starts Pearl gateway,
+  resolves `Qwen3_5ForConditionalGeneration`, selects Pearl kernels, loads all
+  fifteen safetensors shards, exposes `pearl-ai/Qwen3.6-27B-pearl` at
+  `/v1/models`, completes a chat prompt, and passes `jarvis mine
+  validate-model --allow-planned`. Required validation flags:
+  `--language-model-only`, `--skip-mm-profiling`, and
+  `--gdn-prefill-backend triton` with a 4096-token context. Validation
+  artifact:
+  `/tmp/openjarvis-h100/converted/Qwen3.6-27B-pearl-experimental-validate.json`.
+  The published `ScalingIntelligence/Qwen3.6-27B-pearl` repo includes the same
+  validation artifact as `openjarvis_validation.json`.
 
 ## Enablement Checklist
 
@@ -93,7 +123,7 @@ Current validation evidence:
    OpenJarvis includes an experimental local converter for this work:
 
    ```bash
-   python scripts/mining/pearl_model_converter.py \
+   python scripts/pearl/model_converter.py \
      Qwen/Qwen3.5-9B \
      /tmp/pearl-ai-Qwen3.5-9B-pearl \
      --device cuda
