+++ src/openjarvis/mining/_models.py
@@ -72,6 +72,18 @@ PEARL_MODEL_SPECS: tuple[PearlModelSpec, ...] = (
         ),
     ),
     PearlModelSpec(
+        model_id="ScalingIntelligence/Qwen3.6-27B-pearl",
+        base_model_id="Qwen/Qwen3.6-27B",
+        status="validated",
+        min_vram_gb=80.0,
+        default_max_model_len=4096,
+        notes=(
+            "Experimental OpenJarvis/Pearl artifact validated on H100. "
+            "Runtime requires --language-model-only, --skip-mm-profiling, "
+            "and --gdn-prefill-backend triton."
+        ),
+    ),
+    PearlModelSpec(
         model_id="pearl-ai/Gemma-4-E4B-it-pearl",
         base_model_id="google/gemma-4-E4B-it",
         status="planned",
@@ -80,6 +92,17 @@ PEARL_MODEL_SPECS: tuple[PearlModelSpec, ...] = (
         notes=(
             "Planned target; Pearl artifact is not publicly available yet. "
             "Validation tracked in open-jarvis/OpenJarvis#318."
+        ),
+    ),
+    PearlModelSpec(
+        model_id="ScalingIntelligence/Gemma-4-E4B-it-pearl",
+        base_model_id="google/gemma-4-E4B-it",
+        status="validated",
+        min_vram_gb=24.0,
+        default_max_model_len=8192,
+        notes=(
+            "Experimental OpenJarvis/Pearl artifact validated on H100 with "
+            "Gemma4 per-layer embeddings excluded from Pearl quantization."
         ),
     ),
     PearlModelSpec(
