+++ tests/pearl/test_model_converter.py
@@ -15,7 +15,7 @@ save_file = safetensors_torch.save_file
 
 
 def _load_converter():
-    path = Path(__file__).parents[2] / "scripts" / "mining" / "pearl_model_converter.py"
+    path = Path(__file__).parents[2] / "scripts" / "pearl" / "model_converter.py"
     spec = importlib.util.spec_from_file_location("pearl_model_converter", path)
     assert spec is not None and spec.loader is not None
     module = importlib.util.module_from_spec(spec)
@@ -37,6 +37,9 @@ def test_converter_quantizes_linear_weights_and_writes_pearl_config(
     save_file(
         {
             "model.embed_tokens.weight": torch.ones(4, 4, dtype=torch.bfloat16),
+            "model.embed_tokens_per_layer.weight": torch.full(
+                (4, 4), 2, dtype=torch.bfloat16
+            ),
             "model.layers.0.self_attn.q_proj.weight": torch.arange(
                 16, dtype=torch.float32
             ).reshape(4, 4),
@@ -65,6 +68,8 @@ def test_converter_quantizes_linear_weights_and_writes_pearl_config(
     assert tensors["model.layers.0.self_attn.q_proj.weight_scale"].shape == (4, 1)
     assert tensors["model.layers.0.self_attn.o_proj.weight_scale"].shape == (4, 1)
     assert tensors["model.embed_tokens.weight"].dtype == torch.bfloat16
+    assert tensors["model.embed_tokens_per_layer.weight"].dtype == torch.bfloat16
+    assert "model.embed_tokens_per_layer.weight_scale" not in tensors
 
     config = json.loads((out / "config.json").read_text())
     assert config["quantization_config"]["quant_method"] == "pearl"
@@ -73,6 +78,7 @@ def test_converter_quantizes_linear_weights_and_writes_pearl_config(
         == 7
     )
     assert "re:.*visual.*" in config["quantization_config"]["ignore"]
+    assert "re:.*embed_tokens_per_layer$" in config["quantization_config"]["ignore"]
     index = json.loads((out / "model.safetensors.index.json").read_text())
     assert "model.layers.0.self_attn.o_proj.weight_scale" in index["weight_map"]
